pyproject.toml
[project]
name = "ParliamentSearch"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"jupyterlab",
	"transformers>=4.30.0", # Pin a version range compatible with newer ST
	"sentence-transformers>=2.6.0", # Pin a version range that works with newer transformers
	"pandas",
	"pandas",
	"qdrant-client",
	"fastapi",
	"dotenv",
	"psycopg2-binary",
	"uvicorn",
]
